Changed column type to specify NUMBER(10,0) to match Roo-autogenerated datatype for java.Integer

+++ src/main/resources/db/migration/oracle/V1.0.2.3__container_row_column_index_columns.sql
@@ -1,2 +1,2 @@
-ALTER TABLE "ACAS"."CONTAINER" ADD (column_index INTEGER,
-							row_index INTEGER);
+ALTER TABLE "ACAS"."CONTAINER" ADD (column_index NUMBER(10,0),
+							row_index NUMBER(10,0));
